# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Classroom/Venue Management System API",
    "description": "API documentation for the Classroom/Venue Management System.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://wiman.azurewebsites.net/api/",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000/api/",
      "description": "Local server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/venues/{venueId}/issue-report": {
      "post": {
        "tags": ["Venues"],
        "summary": "Create a new issue report",
        "description": "Create a new issue report with an optional image upload.",
        "operationId": "createIssueReport",
        "parameters": [
          {
            "name": "venueId",
            "in": "path",
            "description": "ID of the venue",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "type": "string"
                  },
                  "images": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Issue report created successfully"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/admin/bookings": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get all bookings",
        "description": "Retrieve all bookings in the system.",
        "operationId": "getAllBookings",
        "responses": {
          "200": {
            "description": "List of all bookings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Booking"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "x-swagger-router-controller": "Admin"
      }
    },
...
    "/admin/bookings/approve/{id}": {
      "put": {
        "tags": ["Admin"],
        "summary": "Approve a booking",
        "description": "Approve a pending booking by its ID.",
        "operationId": "confirmBooking",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Booking ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Booking approved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Booking Approved"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Booking not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "x-swagger-router-controller": "Admin"
      }
    },
    "/admin/bookings/cancel/{id}": {
      "put": {
        "tags": ["Admin"],
        "summary": "Cancel a booking",
        "description": "Cancel a booking by its ID.",
        "operationId": "cancelBooking",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Booking ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Booking canceled",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Booking canceled"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Booking not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "x-swagger-router-controller": "Admin"
      }
    },
...
    "/maintenance/issue-reports": {
      "get": {
        "tags": ["Maintenance"],
        "summary": "Get all issue reports",
        "description": "Retrieve a list of all issue reports in the system.",
        "operationId": "getAllIssueReports",
        "responses": {
          "200": {
            "description": "A list of issue reports",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IssueReport"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/maintenance/issue-reports/{id}": {
      "get": {
        "tags": ["Maintenance"],
        "summary": "Get an issue report by ID",
        "description": "Retrieve a specific issue report by its ID.",
        "operationId": "getIssueReportById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The ID of the issue report to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "The requested issue report",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueReport"
                }
              }
            }
          },
          "404": {
            "description": "Issue report not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/maintenance/issue-report/{id}/review": {
      "put": {
        "tags": ["Maintenance"],
        "summary": "Add a review to an issue report",
        "description": "Add a review and update the issue state for a specific issue report.",
        "operationId": "addReviewToIssueReportReview",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The ID of the issue report to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Review"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Review added successfully"
          },
          "404": {
            "description": "Issue report not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/maintenance/issue-report/{id}/add-setback": {
      "put": {
        "tags": ["Maintenance"],
        "summary": "Add a setback to an issue report",
        "description": "Report a setback and update the resolution log of an issue report.",
        "operationId": "addIssueSetBackReport",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The ID of the issue report to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetbackReport"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Setback reported successfully"
          },
          "404": {
            "description": "Issue report not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/maintenance/issue-report/{id}/close": {
      "put": {
        "tags": ["Maintenance"],
        "summary": "Close an issue report",
        "description": "Mark an issue report as closed after resolution.",
        "operationId": "closeIssueReport",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The ID of the issue report to close"
          }
        ],
        "responses": {
          "200": {
            "description": "Issue report closed successfully"
          },
          "404": {
            "description": "Issue report not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "Booking": {
        "type": "object",
        "properties": {
          "bookingId": {
            "type": "integer",
            "description": "Unique identifier for the booking"
          },
          "date": {
            "type": "string",
            "format": "date",
            "description": "Date of the booking (YYYY-MM-DD)"
          },
          "startTime": {
            "type": "string",
            "format": "time",
            "description": "Start time of the booking (HH:MM:SS)"
          },
          "endTime": {
            "type": "string",
            "format": "time",
            "description": "End time of the booking (HH:MM:SS)"
          },
          "venueId": {
            "type": "integer",
            "description": "Venue ID for the booking"
          },
          "status": {
            "type": "string",
            "enum": ["pending", "cancelled", "confirmed"],
            "description": "Status of the booking"
          },
          "reasonForCancellation": {
            "type": "string",
            "description": "Reason for cancellation (if applicable)"
          },
          "repeatFrequency": {
            "type": "string",
            "enum": ["none", "daily", "weekly", "monthly"],
            "description": "Frequency of repeating bookings"
          },
          "repeatUntil": {
            "type": "string",
            "format": "date",
            "description": "Date until which the booking repeats (optional, can be null)"
          },
          "eventName": {
            "type": "string",
            "description": "Name of the event"
          }
        },
        "required": ["id", "date", "startTime", "endTime", "venueId", "status", "eventName"]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Resource not found or invalid operation"
          }
        }
      },
      "IssueReport": {
        "type": "object",
        "properties": {
          "issueId": {
            "type": "integer",
            "description": "The unique identifier of the issue report"
          },
          "venueId": {
            "type": "string",
            "description": "The venue associated with the issue"
          },
          "reportedBy": {
            "type": "string",
            "description": "The user who reported the issue"
          },
          "issueDescription": {
            "type": "string",
            "description": "A detailed description of the issue"
          },
          "reportedDate": {
            "type": "string",
            "format": "date-time",
            "description": "The date and time when the issue was reported"
          },
          "resolvedDate": {
            "type": "string",
            "format": "date-time",
            "description": "The date and time when the issue was resolved",
            "nullable": true
          },
          "resolutionLog": {
            "type": "string",
            "description": "The log of resolutions or actions taken to resolve the issue",
            "nullable": true
          },
          "imageUrl": {
            "type": "string",
            "description": "The URL of an image associated with the issue report",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": ["Reported", "In Progress", "Resolved"],
            "description": "The current status of the issue report"
          }
        }
      },
      "Review": {
        "type": "object",
        "properties": {
          "review": {
            "type": "string",
            "description": "Review comments or further details for the issue"
          },
          "issueState": {
            "type": "string",
            "description": "The current state or classification of the issue"
          }
        }
      },
      "SetbackReport": {
        "type": "object",
        "properties": {
          "setback": {
            "type": "string",
            "description": "Details of the setback in resolving the issue"
          },
          "newFixRequirements": {
            "type": "string",
            "description": "New requirements or changes needed to fix the issue"
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}